Move database altering commands
