+++ src/main/resources/strings.xml
@@ -14,4 +14,5 @@
     <entry key="RAW_MATERIAL">Material</entry>
     <entry key="AMOUNT">Amount</entry>
     <entry key="UPDATED">Updated</entry>
+    <entry key="COMMANDS">Possible commands:</entry>
 </properties>
